requirements.txt
# Core dependencies for Autonomous Field Engineer
boto3>=1.34.0
botocore>=1.34.0

# Data validation and serialization
pydantic>=2.5.0
dataclasses-json>=0.6.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
hypothesis>=6.92.0  # Property-based testing

# Image processing
Pillow>=10.1.0

# Vector database
weaviate-client>=3.26.7,<4.0.0  # Using v3 for compatibility
# Note: sentence-transformers removed — Nova Multimodal Embeddings replaces CLIP

# Web UI
chainlit>=1.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
